chore(test): remove `network_utils` module
